refactor: adjusted server file structure
