+++ clear-ledger-server/src/main/java/com/onixbyte/clearledger/security/UserInterceptor.java
@@ -1,6 +1,6 @@
 package com.onixbyte.clearledger.security;
 
-import com.onixbyte.clearledger.data.biz.BizUser;
+import com.onixbyte.clearledger.data.dto.BizUser;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.security.core.context.SecurityContextHolder;
